Enhance: Button compatible with web-only props
