+++ src/Button/GenericButtonWrapper.js
@@ -3,17 +3,14 @@
 import * as React from 'react';
 
 import Touchable from './Touchable';
-import StyleSheet from '../PlatformStyleSheet';
 import type { StylePropType } from '../PlatformStyleSheet/StyleTypes';
-import type { ButtonType } from './ButtonTypes';
-import { wrapperColor } from './styles';
 
 type Props = {|
   +children: React.Node,
+  +width?: number,
   +onPress: () => void,
   +style?: StylePropType,
   +disabled?: boolean,
-  +type: ButtonType,
   +testID?: string,
 |};
 
@@ -24,8 +21,8 @@ type Props = {|
  */
 export default function Button({
   onPress,
+  width,
   disabled = false,
-  type = 'primary',
   style,
   testID,
   children,
@@ -36,31 +33,13 @@ export default function Button({
     <Touchable
       disabled={!onPress || disabled}
       onPress={onPressHandler}
-      style={[styleSheet.buttonWrapper, theme(type).wrapper, style]}
+      style={style}
       accessibilityComponentType="button"
       accessibilityTraits="button"
       testID={testID}
+      width={width}
     >
       {children}
     </Touchable>
   );
 }
-
-const styleSheet = StyleSheet.create({
-  buttonWrapper: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: 8,
-    paddingVertical: 11,
-    height: 44,
-    borderRadius: 6,
-  },
-});
-
-const theme = (type: ButtonType = 'primary') =>
-  StyleSheet.create({
-    wrapper: {
-      backgroundColor: wrapperColor[type],
-    },
-  });
